requirements.txt
# Core dependencies
python-dotenv>=1.0.0
Pillow>=10.1.0
openai>=1.6.0
tenacity>=8.0.1
aiohttp>=3.9.1
httpx>=0.25.2

# Database
asyncpg>=0.30.0
async_timeout>=4.0.3

# Document processing
PyPDF2>=3.0.1
python-docx>=0.8.11

# Web framework
quart>=0.19.0
quart-cors>=0.7.0
hypercorn>=0.15.0

# Type checking
typing-extensions>=4.9.0
pydantic>=2.5.3

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development tools
pylint>=2.17.0
black>=23.0.0
setuptools>=58.0.4

# Additional dependencies
aiosignal>=1.3.1
annotated-types>=0.6.0
anyio>=4.2.0
async-timeout>=4.0.3
attrs>=23.1.0
blinker>=1.7.0
certifi>=2023.11.17
charset-normalizer>=3.3.2
click>=8.1.7
distro>=1.8.0
frozenlist>=1.4.1
h11>=0.14.0
idna>=3.6
multidict>=6.0.4
priority>=2.0.0
pydantic_core>=2.14.6
sniffio>=1.3.0
tqdm>=4.66.1
wsproto>=1.2.0
yarl>=1.9.4
protobuf==5.26.1
packaging==23.2
